feat(deploy.yml): Adds comments to deploy
In case someone does not read the instructions in getting-started.md
they can see the instructions here in the GitHub action file when the
action fails. I also submitted an issue to the mystmd repository to put
instructions in this file by default for all future users of
`myst init --gh-pages`.